Move get_voltage_electroanatomic to case_routines
