+++ openep/draw/__init__.py
@@ -4,5 +4,4 @@ from .draw_routines import (
     get_freeboundaries,
     draw_free_boundaries,
     draw_map,
-    get_voltage_electroanatomic,
 )
